pyproject.toml
[project]
name = "mlops-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.67.1",
    "fastapi>=0.135.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.11.1",
    "pandas>=2.3.3",
    "pydantic>=2.13.0",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "ruff>=0.15.10",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "uvicorn>=0.44.0",
]
